requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via starlette
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via uvicorn
fastapi==0.95.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
psycopg2-binary==2.9.6
    # via -r requirements.in
pydantic==1.10.7
    # via fastapi
requests==2.29.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
urllib3==1.26.15
    # via requests
uvicorn==0.22.0
    # via -r requirements.in
